requirements/dev.txt
asynctest==0.12.2
flake8==3.6.0
flake8-print==3.1.0
mypy==0.720
pytest-aiohttp==0.3.0
pytest-asyncio==0.9.0
pytest-cov==2.6.0
pytest-mock==1.10.0
pytest==3.8.1
requests-mock==1.6.0
